fix(dashboard): de-dup binding panel, align it, rename fix labels, id-first node picker
- move binding panel out of node-details into matter-endpoint-view as its own
  container (fixes duplication on the binding cluster page + left-edge alignment)
- move getEndpointDeviceTypes to util/endpoints to break the import cycle
- binding ACL fix buttons: 'Fix ACL' / 'Fix → Operate'
- node selection options show node id first, then name
